refactor: add uploader interface
