+++ cmd/server/router.go
@@ -11,9 +11,9 @@ import (
 	"github.com/pet-sitter/pets-next-door-api/cmd/server/handler"
 	"github.com/pet-sitter/pets-next-door-api/internal/configs"
 	"github.com/pet-sitter/pets-next-door-api/internal/domain/auth"
+	s3infra "github.com/pet-sitter/pets-next-door-api/internal/infra/bucket"
 	"github.com/pet-sitter/pets-next-door-api/internal/infra/database"
 	kakaoinfra "github.com/pet-sitter/pets-next-door-api/internal/infra/kakao"
-	s3infra "github.com/pet-sitter/pets-next-door-api/internal/infra/s3"
 	"github.com/pet-sitter/pets-next-door-api/internal/service"
 	pndmiddleware "github.com/pet-sitter/pets-next-door-api/lib/middleware"
 	"github.com/rs/zerolog"
