+++ internal/infra/bucket/client.go
@@ -1,7 +1,12 @@
-package s3infra
+package bucketinfra
 
 import (
 	"io"
+	"path/filepath"
+
+	"github.com/aws/aws-sdk-go/private/protocol/rest"
+	"github.com/google/uuid"
+	pnd "github.com/pet-sitter/pets-next-door-api/api"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
@@ -10,6 +15,10 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/rs/zerolog/log"
 )
+
+type FileUploader interface {
+	UploadFile(file io.ReadSeeker, fileName string) (url string, appError *pnd.AppError)
+}
 
 type S3Client struct {
 	s3Client   *s3.S3
@@ -36,10 +45,27 @@ func NewS3Client(keyID, key, endpoint, region, bucketName string) (*S3Client, er
 	}, nil
 }
 
-func (c *S3Client) UploadFile(file io.ReadSeeker, fileName, _ string) (*s3.PutObjectOutput, error) {
+func (c *S3Client) UploadFile(file io.ReadSeeker, fileName string) (string, *pnd.AppError) {
+	randomFileName := generateRandomFileName(fileName)
+	fullPath := "media/" + randomFileName
+
+	if _, err := c.uploadToS3(file, fullPath); err != nil {
+		return "", pnd.ErrUnknown(err)
+	}
+
+	req, _ := c.GetFileRequest(fullPath)
+	rest.Build(req)
+	if err := req.Send(); err != nil {
+		return "", pnd.ErrUnknown(err)
+	}
+
+	return req.HTTPRequest.URL.String(), nil
+}
+
+func (c *S3Client) uploadToS3(file io.ReadSeeker, fullPath string) (*s3.PutObjectOutput, error) {
 	result, err := c.s3Client.PutObject(&s3.PutObjectInput{
 		Bucket: aws.String(c.bucketName),
-		Key:    aws.String(fileName),
+		Key:    aws.String(fullPath),
 		Body:   file,
 	})
 	if err != nil {
@@ -54,4 +80,9 @@ func (c *S3Client) GetFileRequest(fileName string) (req *request.Request, output
 		Bucket: aws.String(c.bucketName),
 		Key:    aws.String(fileName),
 	})
+}
+
+func generateRandomFileName(originalFileName string) string {
+	extension := filepath.Ext(originalFileName)
+	return uuid.New().String() + extension
 }
